converted file offsets to 64big (long) which fixes out of memory issue encountered with input sources over 2gb jira:ION-61  cr:jonker
[GF:2462384:BrazilPerforce://brazil/src/shared/platform/IonJava/mainline]

+++ src/com/amazon/ion/impl/IonCharacterReader.java
@@ -37,10 +37,10 @@ public class IonCharacterReader extends PushbackReader {
      */
     public static final int BUFFER_PADDING = 1;
 
-    private int m_consumed;
-    private int m_line;
-    private int m_column;
-    private int m_size;
+    private long m_consumed;
+    private int  m_line;
+    private int  m_column;
+    private int  m_size;
 
     // our offset stack--really, for efficiency this should
     // be a linked list of primitive ints to avoid boxing costs
@@ -92,7 +92,7 @@ public class IonCharacterReader extends PushbackReader {
      *
      * @return The logical number of consumed characters.
      */
-    public final int getConsumedAmount() {
+    public final long getConsumedAmount() {
         return m_consumed;
     }
